Rename side filters into page filters
