+++ frontend/integration_test/find_shortcut_test.dart
@@ -8,7 +8,7 @@ import 'package:testcase_dashboard/models/artefact.dart';
 import 'package:testcase_dashboard/models/family_name.dart';
 import 'package:testcase_dashboard/providers/api.dart';
 import 'package:testcase_dashboard/repositories/api_repository.dart';
-import 'package:testcase_dashboard/ui/page_search_bar.dart';
+import 'package:testcase_dashboard/ui/page_filters/page_search_bar.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
